refactor: extract dictation events from App
