+++ src/app/App.tsx
@@ -222,6 +222,8 @@ import { useAppUpdateActions } from "./use-app-update-actions";
 
 import { createNoteActions } from "./note-actions";
 
+import { useDictationEvents } from "./use-dictation-events";
+
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
   const activeHermesProfileName = useActiveHermesProfileName();
@@ -1495,60 +1497,12 @@ export function App() {
     return () => window.removeEventListener(AGENT_GALLERY_EVENT, onGallery);
   }, []);
 
-  useEffect(() => {
-    let aborted = false;
-    let unlisten: (() => void) | undefined;
-    void listen<string>("dictation-event", (event) => {
-      const helperEvent = parseDictationHelperEvent(event.payload);
-      if (!helperEvent) return;
-      dictationWorkflowActiveRef.current = nextDictationWorkflowActive(
-        dictationWorkflowActiveRef.current,
-        helperEvent.type,
-      );
-      if (helperEvent.type === "final_transcript") {
-        // T3 of the referral delight nudge: a dictation landed (often while
-        // June is backgrounded; the card waits to be found).
-        recordDictationFinished();
-        return;
-      }
-      if (helperEvent.type === "agent_session_prompt") {
-        const prompt = stringPayloadValue(helperEvent.payload?.prompt) ?? "";
-        dispatchAgentSessionStatus({
-          prompt,
-          title: titleFromPrompt(prompt),
-          status: "received",
-          summary: "June is starting.",
-        });
-        markAgentNewSessionPending(prompt);
-        setActiveView("agent");
-        window.setTimeout(() => {
-          window.dispatchEvent(
-            new CustomEvent<AgentNewSessionDetail>(AGENT_NEW_SESSION_EVENT, {
-              detail: { prompt },
-            }),
-          );
-        }, 0);
-        return;
-      }
-      if (
-        helperEvent.type !== "permission_status" &&
-        helperEvent.type !== "dictation_diagnostics"
-      ) {
-        return;
-      }
-      const microphone = stringPayloadValue(helperEvent.payload?.microphone);
-      const accessibility = stringPayloadValue(helperEvent.payload?.accessibility);
-      if (microphone) setMicrophoneStatus(microphone);
-      if (accessibility) setAccessibilityStatus(accessibility);
-    }).then((cleanup) => {
-      if (aborted) cleanup();
-      else unlisten = cleanup;
-    });
-    return () => {
-      aborted = true;
-      unlisten?.();
-    };
-  }, []);
+  useDictationEvents({
+    dictationWorkflowActiveRef,
+    setAccessibilityStatus,
+    setActiveView,
+    setMicrophoneStatus,
+  });
 
   // The detached meeting HUD (shown when June is backgrounded, minimized, or
   // hidden mid-recording) is a presence indicator, not a control surface:
